+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -3,6 +3,7 @@
 - Improvement: Added CSharpWhileStatement for adding `while(true) { }` statements.
 - Improvement: It is now possible specify `when` expressions on `CSharpCatchBlock`s.
 - Improvement: Introduced `DefaultIntentManaged` and `IntentTagMode` attributes for File Builder.
+- Improvement: CSharpParameter now has ability to store metadata.
 
 ### Version 3.3.41