+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -6,6 +6,9 @@
 
 - Improvement: Added CSharpWhileStatement for adding `while(true) { }` statements.
 - Improvement: It is now possible specify `when` expressions on `CSharpCatchBlock`s.
+- Improvement: Introduced `DefaultIntentManaged` and `IntentTagMode` attributes for File Builder.
+- Improvement: CSharpParameter now has ability to store metadata.
+- Improvement: CSharpConstructorParameter will propagate metadata to created fields and properties.
 
 ### Version 3.3.41