+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,4 +1,8 @@
-﻿### Version 3.3.42
+﻿### Version 3.3.43
+
+- Improvement: Added support for `await` on `foreach` statements.
+
+### Version 3.3.42
 
 - Improvement: Added CSharpWhileStatement for adding `while(true) { }` statements.
 - Improvement: It is now possible specify `when` expressions on `CSharpCatchBlock`s.